feat(cardwired): add a new Env API to replace switcheroo shim call
