+++ crates/cardwire-daemon/src/core/mod.rs
@@ -1,6 +1,6 @@
+pub mod env;
 pub mod errors;
 pub mod gpu;
+pub mod inode;
 pub mod pci;
 pub mod procfs;
-
-pub mod inode;
